requirements.txt
pandas>=1.5.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0
scikit-learn>=1.0.0
xgboost>=1.5.0
imbalanced-learn>=0.8.0
joblib>=1.1.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-groq>=0.0.1
groq>=0.4.0
python-dotenv>=0.19.0
streamlit>=1.28.0
